pyproject.toml
[tool.poetry]
name = "walle"
version = "0.1.0"
description = ""
authors = ["takuto <pysan3@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
attrdict = "^2.0.1"
rich = "^12.2.0"
toml = "^0.10.2"
Responder-Login = "^0.0.8"
SQLAlchemy = "^1.4.34"
PyJWT = "^2.3.0"
gunicorn = "^20.1.0"
cryptography = "^36.0.2"
mysqlclient = "^2.1.0"
mypy-protobuf = "^3.2.0"
protobuf = "^3.20.0"
hashids = "^1.3.1"
responder = "^2.0.7"
alembic = "^1.7.7"
SQLAlchemy-Utils = "^0.38.2"
typesystem = "<0.3.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
pylint = "^2.13.4"
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
